+++ lightorchestrator/service/pbconvert/device.go
@@ -17,29 +17,29 @@ func NewPBDevice(device device.Device) *pb.Device {
 
 func NewPBVector(vector space.Vector) *pb.Vector {
 	return &pb.Vector{
-		X: vector.X,
-		Y: vector.Y,
-		Z: vector.Z,
+		X: float32(vector.X),
+		Y: float32(vector.Y),
+		Z: float32(vector.Z),
 	}
 }
 
 func NewVector(vector *pb.Vector) space.Vector {
 	return space.Vector{
-		X: vector.X,
-		Y: vector.Y,
-		Z: vector.Z,
+		X: float64(vector.X),
+		Y: float64(vector.Y),
+		Z: float64(vector.Z),
 	}
 }
 
 func NewPBOrientation(orientation space.Orientation) *pb.Orientation {
 	return &pb.Orientation{
-		Theta: orientation.Theta,
-		Phi:   orientation.Phi,
+		Theta: float32(orientation.Theta),
+		Phi:   float32(orientation.Phi),
 	}
 }
 func NewOrientation(orientation *pb.Orientation) space.Orientation {
 	return space.Orientation{
-		Theta: orientation.Theta,
-		Phi:   orientation.Phi,
+		Theta: float64(orientation.Theta),
+		Phi:   float64(orientation.Phi),
 	}
 }
